PairCorrelationFunction

Documentation for PairCorrelationFunction.

PairCorrelationFunction.Constants

PairCorrelationFunction.pcf

PairCorrelationFunction.Constants — Type

Constants

A struct to hold the constants needed for the pair correlation function calculation.

Fields

grid_size::NTuple{N, Float64}: The size of the grid in each dimension.

base_point::NTuple{N, Float64}: The base point of the grid to be used to calculate distance to the boundary.

domain_volume::Float64: The volume of the domain.

radii::AbstractRange{<:Real}: The range of radii to be used for the pair correlation function.

radii2::AbstractVector{<:Real}: The squared values of the radii.

Examples

using PairCorrelationFunction
xlims = (-450.0, 450.0)
ylims = (-450.0, 450.0)
radii = 0:20.0:1300.0
constants = Constants(xlims, ylims, radii)
# output
Constants for 2D pair correlation function:
  grid_size: (900.0, 900.0)
  base_point: (-450.0, -450.0)
  domain_volume: 810000.0
  radii: 0.0 - 1300.0
  #annuli: 65

PairCorrelationFunction.pcf — Method

pcf(centers::AbstractMatrix{<:Real}, targets::AbstractMatrix{<:Real}, constants::Constants)

Calculate the pair correlation function for a set of centers and targets.

For each point in the centers matrix, compute the distance to each point in the targets matrix. Bin these distances by the radii defined in the constants object.

Arguments

centers::AbstractMatrix{<:Real}: A matrix of centers, where each row is a center.

targets::AbstractMatrix{<:Real}: A matrix of targets, where each row is a target.

constants::Constants: A Constants object containing the grid size, base point, domain volume, and radii.

Returns

pcf::Vector{Float64}: A vector of normalized target densities for each annulus around all centers.